pyproject.toml
[project]
name = "superbill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.6",
    "gunicorn>=23.0.0",
    "langchain>=1.2.8",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=0.3.0",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
